app/requirements.txt
blinker==1.6.2
click==8.1.6
Flask==2.3.2
future==0.18.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-whois==0.8.0
Werkzeug==2.3.6
